Separate prject details module, show short IDs, both apps and container
