+++ src/commands/context/get.tsx
@@ -9,58 +9,16 @@ import { Set as SetCommand } from "./set.js";
 import { RenderJson } from "../../rendering/react/json/RenderJson.js";
 import { useRenderContext } from "../../rendering/react/context.js";
 import { LocalFilename } from "../../rendering/react/components/LocalFilename.js";
-import { MittwaldAPIV2 } from "@mittwald/api-client";
-import { assertStatus } from "@mittwald/api-client-commons";
 import Context, {
   ContextKey,
   ContextValue,
   ContextValueSource,
 } from "../../lib/context/Context.js";
 import {
-  getAppFromUuid,
-  getAppInstallationFromUuid,
-} from "../../lib/resources/app/uuid.js";
-
-type AppLinkedDatabase = MittwaldAPIV2.Components.Schemas.AppLinkedDatabase;
-
-type LinkedDatabaseSummary = {
-  databaseId: string;
-  purpose: string;
-  kind: "mysql" | "redis" | "unknown";
-  name?: string;
-};
-
-type AppSummary = {
-  installationId: string;
-  appId: string;
-  appName: string;
-  installationPath: string;
-  linkedDatabases: LinkedDatabaseSummary[];
-};
-
-type StackSummary = {
-  id: string;
-  description?: string;
-  services: number;
-  volumes: number;
-};
-
-type ContainerSummary = {
-  id: string;
-  name: string;
-  status: string;
-  stackId?: string;
-};
-
-type ProjectOverview = {
-  projectId?: string;
-  projectName?: string;
-  resolvedFrom?: "project-id" | "installation-id";
-  apps: AppSummary[];
-  stacks: StackSummary[];
-  containers: ContainerSummary[];
-  unavailableReason?: string;
-};
+  fetchProjectOverview,
+  ProjectOverview,
+  resolveProjectContext,
+} from "../../lib/context/projectOverview.js";
 
 const ContextSourceValue: FC<{ source: ContextValueSource }> = ({ source }) => {
   switch (source.type) {
@@ -125,6 +83,10 @@ const ContextSource: FC<{ source: ContextValueSource }> = ({ source }) => {
 const ProjectOverviewSection: FC<{ overview: ProjectOverview }> = ({
   overview,
 }) => {
+  const stackDisplayById = new Map(
+    overview.stacks.map((stack) => [stack.id, stack.shortId ?? stack.id]),
+  );
+
   if (overview.unavailableReason) {
     return (
       <Note marginBottom={1}>
@@ -137,21 +99,24 @@ const ProjectOverviewSection: FC<{ overview: ProjectOverview }> = ({
     Project: (
       <Text>
         <Value>{overview.projectName ?? overview.projectId}</Value>{" "}
-        <Text color="gray">({overview.projectId})</Text>
+        <Text color="gray">
+          ({overview.projectShortId ?? overview.projectId})
+        </Text>
       </Text>
     ),
     "Resolved from": <Value>{overview.resolvedFrom ?? "project-id"}</Value>,
   };
 
-  if (overview.apps.length > 0) {
-    rows["Apps"] = (
+  rows["Apps"] =
+    overview.apps.length > 0 ? (
       <Box flexDirection="column">
         {overview.apps.map((app) => (
           <Box key={app.installationId} flexDirection="column" marginBottom={1}>
             <Text>
               <Value>{app.appName}</Value>{" "}
               <Text color="gray">
-                ({app.installationPath}, {app.installationId})
+                ({app.installationPath},{" "}
+                {app.installationShortId ?? app.installationId})
               </Text>
             </Text>
             {app.linkedDatabases.length > 0 ? (
@@ -169,36 +134,49 @@ const ProjectOverviewSection: FC<{ overview: ProjectOverview }> = ({
           </Box>
         ))}
       </Box>
+    ) : (
+      <Text color="gray">none found in this project</Text>
     );
-  } else {
-    rows["Apps"] = <Text color="gray">none found in this project</Text>;
-    rows["Stacks"] = (
+
+  rows["Stacks"] =
+    overview.stacks.length > 0 ? (
       <Box flexDirection="column">
         <Text>
           <Value>{overview.stacks.length}</Value> total
         </Text>
         {overview.stacks.slice(0, 5).map((stack) => (
           <Text key={stack.id} color="gray">
-            {stack.id}: {stack.services} services, {stack.volumes} volumes
+            {stack.shortId ?? stack.id}: {stack.services} services,{" "}
+            {stack.volumes} volumes
             {stack.description ? ` (${stack.description})` : ""}
           </Text>
         ))}
       </Box>
+    ) : (
+      <Text color="gray">none found in this project</Text>
     );
-    rows["Containers"] = (
+
+  rows["Containers"] =
+    overview.containers.length > 0 ? (
       <Box flexDirection="column">
         <Text>
           <Value>{overview.containers.length}</Value> total
         </Text>
         {overview.containers.slice(0, 8).map((container) => (
           <Text key={container.id} color="gray">
-            {container.name}: {container.status}
-            {container.stackId ? ` (stack ${container.stackId})` : ""}
+            {container.shortId
+              ? `${container.shortId} (${container.name})`
+              : container.name}
+            : {container.status}
+            {container.stackId
+              ? ` (stack ${stackDisplayById.get(container.stackId) ?? container.stackId})`
+              : ""}
           </Text>
         ))}
       </Box>
+    ) : (
+      <Text color="gray">none found in this project</Text>
     );
-  }
 
   return <SingleResult title="Project context overview" rows={rows} />;
 };
@@ -241,198 +219,21 @@ const GetContext: FC<{ ctx: Context }> = ({ ctx }) => {
   const appInstallationId = values["installation-id"]?.value;
 
   const resolvedProject = usePromise(
-    async (
+    (
       contextProjectId: string | undefined,
       installationId: string | undefined,
-    ): Promise<{
-      projectId?: string;
-      resolvedFrom?: "project-id" | "installation-id";
-      unavailableReason?: string;
-    }> => {
-      if (contextProjectId) {
-        return { projectId: contextProjectId, resolvedFrom: "project-id" };
-      }
-
-      if (!installationId) {
-        return {
-          unavailableReason:
-            "no project-id in context and no installation-id to derive it from",
-        };
-      }
-
-      try {
-        const installation = await getAppInstallationFromUuid(
-          apiClient,
-          installationId,
-        );
-        return {
-          projectId: installation.projectId,
-          resolvedFrom: "installation-id",
-        };
-      } catch {
-        return {
-          unavailableReason: "could not resolve project from installation-id",
-        };
-      }
-    },
+    ) => resolveProjectContext(apiClient, contextProjectId, installationId),
     [projectIdFromContext, appInstallationId],
   );
 
   const overview = usePromise(
-    async (
-      projectId: string | undefined,
-      resolvedFrom: "project-id" | "installation-id" | undefined,
-      unavailableReason: string | undefined,
-    ): Promise<ProjectOverview> => {
-      if (!projectId) {
-        return {
-          apps: [],
-          stacks: [],
-          containers: [],
-          unavailableReason:
-            unavailableReason ?? "project could not be resolved",
-        };
-      }
-
-      try {
-        const projectResponse = await apiClient.project.getProject({
-          projectId,
-        });
-        assertStatus(projectResponse, 200);
-
-        const appInstallationsResponse =
-          await apiClient.app.listAppinstallations({ projectId });
-        assertStatus(appInstallationsResponse, 200);
-
-        const appInstallations = appInstallationsResponse.data;
-        const uniqueAppIds = Array.from(
-          new Set(appInstallations.map((installation) => installation.appId)),
-        );
-
-        const appNames = new Map<string, string>();
-        await Promise.all(
-          uniqueAppIds.map(async (appId) => {
-            try {
-              const app = await getAppFromUuid(apiClient, appId);
-              appNames.set(appId, app.name);
-            } catch {
-              appNames.set(appId, appId);
-            }
-          }),
-        );
-
-        const databaseById = new Map<
-          string,
-          { name: string; kind: "mysql" | "redis" }
-        >();
-
-        try {
-          const mysqlResponse = await apiClient.database.listMysqlDatabases({
-            projectId,
-          });
-          assertStatus(mysqlResponse, 200);
-          for (const db of mysqlResponse.data) {
-            databaseById.set(db.id, { name: db.name, kind: "mysql" });
-          }
-        } catch {
-          // best effort
-        }
-
-        try {
-          const redisResponse = await apiClient.database.listRedisDatabases({
-            projectId,
-          });
-          assertStatus(redisResponse, 200);
-          for (const db of redisResponse.data) {
-            databaseById.set(db.id, { name: db.name, kind: "redis" });
-          }
-        } catch {
-          // best effort
-        }
-
-        const apps: AppSummary[] = appInstallations.map((installation) => {
-          const linkedDatabases: LinkedDatabaseSummary[] =
-            installation.linkedDatabases.map((linked: AppLinkedDatabase) => {
-              const resolved = databaseById.get(linked.databaseId);
-              return {
-                databaseId: linked.databaseId,
-                purpose: linked.purpose,
-                kind: resolved?.kind ?? "unknown",
-                name: resolved?.name,
-              };
-            });
-
-          return {
-            installationId: installation.id,
-            appId: installation.appId,
-            appName: appNames.get(installation.appId) ?? installation.appId,
-            installationPath: installation.installationPath,
-            linkedDatabases,
-          };
-        });
-
-        if (apps.length > 0) {
-          return {
-            projectId,
-            projectName: projectResponse.data.description,
-            resolvedFrom,
-            apps,
-            stacks: [],
-            containers: [],
-          };
-        }
-
-        const stackResponse = await apiClient.container.listStacks({
-          projectId,
-        });
-        assertStatus(stackResponse, 200);
-
-        const serviceResponse = await apiClient.container.listServices({
-          projectId,
-        });
-        assertStatus(serviceResponse, 200);
-
-        const stacks: StackSummary[] = stackResponse.data.map((stack) => ({
-          id: stack.id,
-          description: stack.description,
-          services: stack.services?.length ?? 0,
-          volumes: stack.volumes?.length ?? 0,
-        }));
-
-        const containers: ContainerSummary[] = serviceResponse.data.map(
-          (service) => ({
-            id: service.id,
-            name: service.serviceName,
-            status: service.status,
-            stackId: service.stackId,
-          }),
-        );
-
-        return {
-          projectId,
-          projectName: projectResponse.data.description,
-          resolvedFrom,
-          apps,
-          stacks,
-          containers,
-        };
-      } catch {
-        return {
-          projectId,
-          resolvedFrom,
-          apps: [],
-          stacks: [],
-          containers: [],
-          unavailableReason:
-            "project-level data could not be fetched with current access/context",
-        };
-      }
-    },
-    [
-      resolvedProject.projectId,
-      resolvedProject.resolvedFrom,
-      resolvedProject.unavailableReason,
-    ],
+    (resolvedProjectContext: {
+      projectId?: string;
+      resolvedFrom?: "project-id" | "installation-id";
+      unavailableReason?: string;
+    }): Promise<ProjectOverview> =>
+      fetchProjectOverview(apiClient, resolvedProjectContext),
+    [resolvedProject],
   );
 
   if (renderAsJson) {
